requirements.txt
fastapi
uvicorn
tensorflow
pillow
numpy
nltk